test: BlueprintServiceTest에서 카테고리 관련 테스트를 CategoryServiceTest로 이동
- BlueprintServiceTest에서 도면 관련 테스트와 관심사가 다르다고 판단
- 이를 해결하기 위해 새로운 CategoryServiceTest 생성
